mongodb sub-pipeline for single-column aggregates
