+++ crates/mongodb-agent-common/src/aggregation_function.rs
@@ -1,10 +1,8 @@
 use enum_iterator::{all, Sequence};
 
-// TODO: How can we unify this with the Accumulator type in the mongodb module?
 #[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, Sequence)]
 pub enum AggregationFunction {
     Avg,
-    Count,
     Min,
     Max,
     Sum,
@@ -17,7 +15,6 @@ impl AggregationFunction {
     pub fn graphql_name(self) -> &'static str {
         match self {
             A::Avg => "avg",
-            A::Count => "count",
             A::Min => "min",
             A::Max => "max",
             A::Sum => "sum",
@@ -30,15 +27,5 @@ impl AggregationFunction {
             .ok_or(QueryPlanError::UnknownAggregateFunction {
                 aggregate_function: s.to_owned().into(),
             })
-    }
-
-    pub fn is_count(self) -> bool {
-        match self {
-            A::Avg => false,
-            A::Count => true,
-            A::Min => false,
-            A::Max => false,
-            A::Sum => false,
-        }
     }
 }
