+++ crates/mongodb-agent-common/src/mongo_query_plan/mod.rs
@@ -116,6 +116,8 @@ pub type ComparisonValue = ndc_query_plan::ComparisonValue<MongoConfiguration>;
 pub type ExistsInCollection = ndc_query_plan::ExistsInCollection<MongoConfiguration>;
 pub type Expression = ndc_query_plan::Expression<MongoConfiguration>;
 pub type Field = ndc_query_plan::Field<MongoConfiguration>;
+pub type Dimension = ndc_query_plan::Dimension<MongoConfiguration>;
+pub type Grouping = ndc_query_plan::Grouping<MongoConfiguration>;
 pub type MutationOperation = ndc_query_plan::MutationOperation<MongoConfiguration>;
 pub type MutationPlan = ndc_query_plan::MutationPlan<MongoConfiguration>;
 pub type MutationProcedureArgument = ndc_query_plan::MutationProcedureArgument<MongoConfiguration>;
